ref(AuthService): move User related fns to UserService
